requirements.txt
# Core Dependencies
aiohttp==3.9.1
beautifulsoup4==4.12.2
faiss-cpu==1.7.4
google-generativeai==0.3.1
pandas==2.1.3
numpy>=1.24.3
scrapy==2.11.0

# NLP & AI
sentence-transformers==2.2.2
textblob==0.17.1
torch>=2.0.0
transformers>=4.30.0
networkx>=3.1  # For GNN implementations

# SEO & Web Analysis
python-dotenv==1.0.0
selenium==4.15.2
readability-metrics==1.4.5
google-search-results==2.4.2
waybackpy>=3.0.6
requests>=2.31.0

# Visualization & Reporting
plotly==5.18.0
graphviz>=0.20.1
dash>=2.9.3

# Performance & Monitoring
tqdm>=4.65.0
PyYAML>=6.0.1
prometheus-client>=0.17.0
redis>=4.5.4  # For caching

# Testing & Development
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.3.0